Update mutation for Item
In this commit I added an update mutation for item, so it's not possible to update a specific item based on a given Id. Using the HTTP request of Patch, I only have to change what I want or need to without bothering to send all of the item data in the mutation.